Extract encoding schema parsing logic into its own component
