+++ CMakeLists.txt
@@ -40,6 +40,7 @@ add_subdirectory(vendor/jsontoolkit)
 add_subdirectory(src/alterschema)
 add_subdirectory(src/canonicalizer)
 add_subdirectory(src/encoding)
+add_subdirectory(src/parser)
 add_subdirectory(src/numeric)
 add_subdirectory(src/encoder)
 add_subdirectory(src/decoder)
@@ -58,6 +59,7 @@ if(JSONBINPACK_TESTS)
   add_subdirectory(test/alterschema)
   add_subdirectory(test/canonicalizer)
   add_subdirectory(test/encoding)
+  add_subdirectory(test/parser)
   add_subdirectory(test/numeric)
   add_subdirectory(test/encoder)
   add_subdirectory(test/decoder)
