+++ src/parser/parser.cc
@@ -1,19 +1,15 @@
-#include <jsonbinpack/encoding/encoding.h>
-#include <jsonbinpack/mapper/encoding.h>
-#include <jsonbinpack/mapper/parser.h>
+#include <jsonbinpack/parser/parser.h>
 
-#include <cassert>   // assert
 #include <sstream>   // std::ostringstream
 #include <stdexcept> // std::runtime_error
 #include <string>    // std::string
 
-namespace sourcemeta::jsonbinpack::mapper {
+namespace sourcemeta::jsonbinpack {
 
 auto parse(const sourcemeta::jsontoolkit::Value &input) -> Encoding {
-  assert(is_encoding(input));
   const auto encoding{sourcemeta::jsontoolkit::to_string(
-      sourcemeta::jsontoolkit::at(input, keywords::name))};
-  const auto &options{sourcemeta::jsontoolkit::at(input, keywords::options)};
+      sourcemeta::jsontoolkit::at(input, "name"))};
+  const auto &options{sourcemeta::jsontoolkit::at(input, "options")};
   if (encoding == "CONST_NONE") {
     assert(sourcemeta::jsontoolkit::defines(options, "value"));
     return CONST_NONE(sourcemeta::jsontoolkit::at(options, "value"));
@@ -24,4 +20,4 @@ auto parse(const sourcemeta::jsontoolkit::Value &input) -> Encoding {
   }
 }
 
-} // namespace sourcemeta::jsonbinpack::mapper
+} // namespace sourcemeta::jsonbinpack
