+++ test/mapper/CMakeLists.txt
@@ -2,7 +2,6 @@ add_executable(sourcemeta_jsonbinpack_mapper_unit
   "${PROJECT_SOURCE_DIR}/vendor/googletest/googletest/src/gtest_main.cc"
   is_encoding_unit.cc make_encoding_unit.cc
   mapper_test.cc
-  v1_enum_parser_test.cc
   2020_12_mapper_enum_test.cc
   2020_12_mapper_integer_test.cc
   2020_12_mapper_number_test.cc
@@ -10,6 +9,5 @@ add_executable(sourcemeta_jsonbinpack_mapper_unit
 target_link_libraries(sourcemeta_jsonbinpack_mapper_unit PRIVATE gtest)
 target_link_libraries(sourcemeta_jsonbinpack_mapper_unit PRIVATE sourcemeta_jsonbinpack_mapper)
 target_link_libraries(sourcemeta_jsonbinpack_mapper_unit PRIVATE sourcemeta_jsonbinpack_canonicalizer)
-target_link_libraries(sourcemeta_jsonbinpack_mapper_unit PRIVATE sourcemeta_jsonbinpack_encoding)
 target_link_libraries(sourcemeta_jsonbinpack_mapper_unit PRIVATE sourcemeta_jsontoolkit_json)
 gtest_discover_tests(sourcemeta_jsonbinpack_mapper_unit)
